requirements.txt
selenium==4.27.1
pytest~=8.3.4
trio~=0.28.0
pytest-trio
pytest-rerunfailures
flake8~=7.1.1
requests~=2.32.3
allure-pytest~=2.13.5
allure-python-commons~=2.13.5
urllib3~=2.3.0
webdriver-manager~=4.0.2
pytest-html~=4.1.1
Appium-Python-Client~=4.4.0
Faker~=33.1.0
h11~=0.14.0
pip~=24.3.1
configparser~=7.1.0
attrs~=24.3.0
PySocks~=1.7.1
typing_extensions~=4.12.2
outcome~=1.3.0.post0
sniffio~=1.3.1
sortedcontainers~=2.4.0
idna~=3.10
pyflakes~=3.2.0
pycodestyle~=2.12.1
wcwidth~=0.2.13
pluggy~=1.5.0
iniconfig~=2.0.0
py~=1.11.0
certifi~=2024.12.14
charset-normalizer~=3.4.1
chardet~=5.2.0
trio-websocket~=0.11.1
websocket-client~=1.8.0
packaging~=24.2
wsproto~=1.2.0